+++ includes/cls/database/upgrade/(db_name)/v.8.1_transactionitems.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `transactionitems` (
 `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
 `title` varchar(500) NOT NULL,
 `caller` varchar(500) NOT NULL,
-`unit_id` int(10) unsigned NOT NULL,
+`unit_id` smallint(5) unsigned NOT NULL,
 `type` ENUM('real','gift','prize','transfer') NOT NULL,
 `minus` float(53) NULL,
 `plus` float(53) NULL,
